@@ -20,21 +20,15 @@ final RegExp _lparen = RegExp(r'\(');
 final RegExp _partial = RegExp(r'renderPartial[\s\t]*');
 
 class DirkAST {
-  final String start;
   final String contents;
   final bool isPartial;
   List<DirkError> errors = [];
   List<Token> tree = [];
 
   DirkAST({
-    String start = '',
-    String layout = '@renderBody()',
-    String contents,
+    this.contents = '',
     this.isPartial = false,
-  })  : this.start = start,
-        this.contents = isPartial
-            ? contents
-            : "${start.length > 0 ? start + '\n' : ''}${layout.replaceAll(RegExp(r'@renderBody\(\)'), contents)}";
+  });
 
   String _scanTillEndOfStatement(SpanScanner scanner) {
     var buffer = '';
@@ -126,9 +120,9 @@ class DirkAST {
         //matches: if (...) {...}
         if (_scanner.scan(_if)) {
           if (_scanner.scan(_ifCondition)) {
-            var condition = _scanner.lastMatch.group(1);
+            var condition = _scanner.lastMatch?.group(1);
             var body = _scanTillClosingCurlyBrace(_scanner);
-            tree.add(Token(TokenType.$if, condition));
+            tree.add(Token(TokenType.$if, condition ?? ""));
             _deepParse(tree.last.children, body);
           } else {
             errors.add(DirkError(
@@ -139,9 +133,9 @@ class DirkAST {
         //matches: for(...) {...}
         if (_scanner.scan(_for)) {
           if (_scanner.scan(_forCondition)) {
-            var condition = _scanner.lastMatch.group(1);
+            var condition = _scanner.lastMatch?.group(1);
             var body = _scanTillClosingCurlyBrace(_scanner);
-            tree.add(Token(TokenType.$for, condition));
+            tree.add(Token(TokenType.$for, condition ?? ""));
             _deepParse(tree.last.children, body);
           } else {
             errors.add(DirkError(
@@ -174,12 +168,12 @@ class DirkAST {
         }
         //matches: anyExpression
         if (_scanner.scan(_expr)) {
-          var body = _scanner.lastMatch.group(1);
-          tree.add(Token(TokenType.expression, body));
+          var body = _scanner.lastMatch?.group(1);
+          tree.add(Token(TokenType.expression, body ?? ""));
           while (!_scanner.isDone) {
             //.member match
             if (_scanner.scan(_member)) {
-              tree.last.content += _scanner.lastMatch.group(1);
+              tree.last.content += _scanner.lastMatch?.group(1) ?? "";
               continue;
             }
             //( match
@@ -221,19 +215,19 @@ class DirkAST {
 
   //Compile the tree as a view function
   String toViewFunction(String name) {
+    final isLayoutView = name == "LayoutView";
+
     StringBuffer result = StringBuffer();
     List<Token> imports =
         tree.where((el) => el.type == TokenType.$import).toList();
-    Token model =
-        tree.firstWhere((el) => el.type == TokenType.model, orElse: () => null);
+    Token model = tree.firstWhere((el) => el.type == TokenType.model,
+        orElse: () => Token(TokenType.model, ""));
     tree.removeWhere(
         (el) => el.type == TokenType.$import || el.type == TokenType.model);
-    String modelType = model?.content ?? "";
 
-    result.writeln("// Generated by DirkAST, don't modify");
-    result.writeln("");
-    result.writeln("// ignore_for_file: unused_import, unused_local_variable");
-    result.writeln("import 'views.dirk.dart';");
+    //force String param for LayoutView
+    String modelType = isLayoutView ? "String" : model.content;
+
     imports.forEach((el) => result.writeln(el));
     result.writeln(
         "String ${name}(${modelType != '' ? modelType + ' model' : ''}) {");
@@ -245,9 +239,15 @@ class DirkAST {
 
     tree.forEach((el) => result.writeln(el));
 
-    result.writeln("return res;");
+    result.writeln(
+        isLayoutView || isPartial ? "return res;" : "return LayoutView(res);");
     result.writeln("}");
 
-    return tryFormatCode(result.toString());
+    //for LayoutView replace placeholder with model param
+    var res = isLayoutView
+        ? result.toString().replaceAll(RegExp(r'renderBody\(\)'), 'model')
+        : result.toString();
+
+    return tryFormatCode(res);
   }
 }
